fix(INSIGHT-1108): RC needs no post release commit

+++ tools/release.sh
@@ -50,18 +50,20 @@ else
   jx step changelog -v "v$RELEASE_VERSION"
 fi
 
-# Not including the release tag in master history
-git reset --hard "origin/$BRANCH"
+if [ "$INCREMENT" != 'release' ]; then
+  # Not including the release tag in master history
+  git reset --hard "origin/$BRANCH"
 
-mvn -B versions:set versions:set-property -DnewVersion="${NEXT_VERSION}-SNAPSHOT" -Dproperty=nuxeo.ai.version -DgenerateBackupPoms=false
-sed -i "s,version: .*,version: ${NEXT_VERSION}-SNAPSHOT," charts/*/Chart.yaml
-git add -u
-jx step next-version --version="$NEXT_VERSION"
-git commit -a -m"Post release ${RELEASE_VERSION} set version ${NEXT_VERSION}-SNAPSHOT"
-if [ "$DRY_RUN" = 'true' ]; then
-  echo "Dry run: skip 'git push' and 'jx start pipeline'"
-  git show
-else
-  git push origin "$BRANCH"
+  mvn -B versions:set versions:set-property -DnewVersion="${NEXT_VERSION}-SNAPSHOT" -Dproperty=nuxeo.ai.version -DgenerateBackupPoms=false
+  sed -i "s,version: .*,version: ${NEXT_VERSION}-SNAPSHOT," charts/*/Chart.yaml
+  git add -u
+  jx step next-version --version="$NEXT_VERSION"
+  git commit -a -m"Post release ${RELEASE_VERSION} set version ${NEXT_VERSION}-SNAPSHOT"
+  if [ "$DRY_RUN" = 'true' ]; then
+    echo "Dry run: skip 'git push' and 'jx start pipeline'"
+    git show
+  else
+    git push origin "$BRANCH"
+  fi
 fi
 chmod +r VERSION charts/*/templates/release.yaml || true
